Added recommender class using singleton and strategy design patterns
Simple models trained on the slim dataset and saved as .pkl files that can subsequently be used by the Recommender class. The class uses the singleton patter as well as model caching such that we don't load the same model more than once.  The strategy pattern allows easy switch between models. The abstract base class dictates the signature of the recommendation that we expect a recommender to be able to make. Errors and logging are used for incorrect model loading or if there is no model loaded at all.